feat: adds creating comments with new depth indicators

+++ lib/validation/models.go
@@ -59,7 +59,7 @@ type CreateComment struct {
 	// [required] the post this comment is associated with
 	PostID uint `json:"post_id" validate:"required"`
 	// the comment this comment is threaded under. Left empty to indicate this is a "root-level" comment
-	ParentCommentID *int64 `json:"parent_comment_id"`
+	ParentCommentID *uint `json:"parent_comment_id"`
 	// [required] the actual text content of the comment
 	Content string `json:"content" validate:"required,min=1,max=500" gorm:"not null"`
 }
